Connect frontend web app to Python AI agent backend (#4)
With this commit, the python AI/LLM/literature parsing code is connected with the frontend React app. The way it works is as follows:

On the main page, user can input a search query for any gene with optional search parameters (max results, top k results returned, etc.). Then the frontend will get a job ID and continously poll the backend for the status of that job.

There will be an attached progress bar indicating to the user how many papers have been processed and how many remain. The user can cancel this search at any time.

Once the paper screening is done, each paper is summarized in a box which is displayed to the user.